requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.2
black==25.9.0
boto3==1.40.50
botocore==1.40.50
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
chromadb-client==1.1.1
click==8.3.0
collection==0.1.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
effdet==0.4.1
emoji==2.15.0
et_xmlfile==2.0.0
executing==2.2.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
google-api-core==2.25.2
google-auth==2.41.1
google-cloud-vision==3.10.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
hf-xet==1.1.10
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lance-namespace==0.0.6
lance-namespace-urllib3-client==0.0.17
lancedb==0.25.1
langchain==0.3.27
langchain-core==0.3.78
langchain-text-splitters==0.3.11
langdetect==1.0.9
langsmith==0.4.32
lxml==6.0.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
msoffcrypto-tool==5.4.2
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
olefile==0.47
omega-conf==2.3.0
onnx==1.19.0
onnxruntime==1.23.1
openai
opencv-python==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20250506
pexpect==4.9.0
pi_heif==1.1.1
pikepdf==9.11.0
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
posthog==5.4.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycocotools==2.0.10
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
Pygments==2.19.2
pylance==0.38.1
pypandoc==1.15
pyparsing==3.2.5
pypdf==6.1.1
pypdfium2==4.30.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv
python-docx==1.2.0
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
RapidFuzz==3.14.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
timm==1.0.20
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
triton==3.4.0
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.15
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.9
zipp==3.23.0
zstandard==0.25.0